7.15 and 7.16 busiest and quietest hours
Added busiestHour() and quiestestHour() methods.
Both use a for loop to check every entry in hourCounts array and compare it to the current highest or lowest value. If the value at i is higher than the current higest value, the variable highest is replaced with the greater value, and the current index value is stored in a variable.
The same happens with quietestHour() but with the lower number.