7.18 Busiest two-hour period
Created busiestTwoHour() method the works about the same as the busiestHour() and quietestHour() methods.
The method tests to see if the sum of the hour count at i and i+1 is greater than the current highest value. If it is, the highest value is reassigned and the hour at i is strored in a variable.
The for loop only goes until i is 2 less than the length of the array to avoid an error when the loop gets to the last entry (hour 23).
I added a final if statement outside the for loop to test the time range 11pm to 1am.