7.19c quietestDay and busiestDay
Created quietestDay and busiestDay methods.
Ran into the problem that while there is an hour 0, there is no day 0, so I increased the size of the arrays by one for day, month, and year. I figured it would be easier to just ignore index 0 then to have to subtract one everytime from the index to get the day I wanted.
I created an analyzeDailyData() method to make the quietestDay and busiestDay possible. It follows the same format as analyzeHourlyData.
Added printDailyCounts() method so I can make sure methods are working correctly.