7.19 Intermediate Step
This commit is in awkward limbo.
I'm running into the problem that since only one reader object is created (in the constructor) once you call one of the analyzeData methods, calling another one doesn't work because the iterator doesn't have a next item anymore so the method just skips the while loop.
I'm debating between taking out the reader object assignment from the constructors and putting it in each of the analyze methods so a new one is created at each method call, or making one analyzeData() method that assigns the all the fields to their corresponding arrays in one fell swoop.