7.19 Complete
Okay, I went the route of creating a new reader object in each analyzeData method so I didn't completely clobber the code structure/ look.
The quietestMonth() and busiestMonth() were added one commit back, but are finalized here and they follow the same framework as the other busiest methods.
Created averageAccessesPerMonth method one commit back as well, but it's finalized here. This method creates a new array for monthly averages and iterates over the log file and at each iteration checks the year and updates the year variable and increases the year counter if valid. Then the total number of accesses for a month is divided by the number of years and that number is assigned to each corresponding month index in monthlyAverages[].
I'm not super happy with the way this code works, due to some weird overlaps. averageAccessesPerMonth() reassigns the monthCounts[] data (making analyzeMonthlyData obsolete), but if I take that part of the code out, then anlayzeMonthlyData() must be called before averageAccessesPerMonth () can be.
Creating a new reader object at each analyzeData method call also generates a bunch of garabage. And with the code the way it currently is, if you accidently call the same analyzeData method twice you corrupt the data and have to recreate a new logAnalyzer object.
I might do the "analyze all data in one method" approach, just to have it look/ work better.