The new and improved LogAnalyzer!
Week 8 - Weblog Analyzer Lab Complete.
Welcome to LogAnalyzerPro! It's basically LogAnalyzer, but better.
Okay, I created a new class that analyzes all the data in the file and adds it to corresponding arrays in one method.
I also fixed some of the documentation that I missed the first time around.
I added the local boolean "analyzed" to avoid corrupting the arrays with multiple calls to the analyzeData() method. Used that boolean to make sure data is analyzed before anyother method is called.
I like this version A LOT better! It looks better and cleaner, and I feel like there's less pitfalls that could break the program.