Extra Credit: accessCodeCount()
Modified LogEntry to have a 6th field for code and added accessor method for CODE.
Modified LogfileCreator to initialize the variable "code" with a random value between 1 - 100. The code is reassigned with actual code (200, 404, or 403) depending on where it fell in the range.
Modified LogfileReader format vaiable to include Code.
I used LogAnalyzerPro for the extra credit. I added a HashMap for the years. The key is the year, and the value is the number of accesses in that year. I added a seperate logfileReader object to use for the extra credit.
Created accessCodeCount() method, which is the extra credit. The method uses reader2 to iterate over the logfile. At each iteration, the entries year is checked against the desired parameter year; if it doesn't match the iterator moves on to the next item. If the year does match, the program starts at Month 01 and checks for a match there. If the months are the same, the corresponding code counter increases by one, and the iterator moves to the next item. If the month doesn't match, that means that there are no entries (or no more entries) for that month in that year and the program prints out the current months info and increases the month counter by one, and rechecks the entry.
If the while loop exits before the month counter reaches 12, the remaining months details are printed.